requirements.txt
Flask
Flask-Cors
Flask-Socketio
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Script==2.0.5
Eventlet==0.30.2
Gunicorn==20.1.0
pytz==2021.3
psycopg2-binary==2.8.5